pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nimotion"
version = "0.1.0"
description = "NiMotion 一体化步进电机调试工具"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
    "PyQt5>=5.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-qt>=4.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
nimotion = "nimotion.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/nimotion"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
